requirements.txt
# MCP Orchestrator dependencies

# LangChain 1.x agent stack (create_agent + LangGraph)
langchain>=1.0.0,<2
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# OpenAI SDK (shared with the MCP-Router)
openai>=1.40.0

# FastAPI service
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# HTTP client (for calling the router and SSE MCP servers)
httpx>=0.27.0

# Config and data
pydantic>=2.5.0
pyyaml>=6.0.1
